requirements.txt
pandas==1.4.2
numpy==1.21.5
yfinance==0.1.74
streamlit==1.16.0
plotly-express==0.4.1
pysqlite3==0.4.7
